Initial input files for terminal basin example
I damned the exit of sagehen and set a single SWALE point a few cells in from the edge of the watershed.  If this example runs through then I will try adding a small lake inclosing the swale point (and set all the lake cell elevations to the same value).