Fixes for terminal lake / terminal basin
Bug fix in stream delination. Arc 10.6 seems to be following the documentation exactly and StreamLink must be an array of 1's and nodata (instead of values > 0 and nodata).
Bug fix in point to raster conversion in dem_parameters needs to be to projected DEM grid
Bug fix when nlakes is set to calculated but there aren't any lakes.